+++ src/simlib/PMCLens.cpp
@@ -5,12 +5,12 @@
 /* The following is a default implementation for the point mass lens. 
  * It would be better to make the class abstract and move this definition to the 
  * subclass. */
-cv::Point2f PointMassLens::getDistortedPos(double r, double theta) const {
+cv::Point2d PointMassLens::getDistortedPos(double r, double theta) const {
     double R = getNuAbs() * CHI ;
     double frac = (einsteinR * einsteinR * r) / (r * r + R * R + 2 * r * R * cos(theta));
     double nu1 = r*cos(theta) + frac * (r / R + cos(theta)) ;
     double nu2 = r*sin(theta) - frac * sin(theta) ;
-    return cv::Point2f( nu1/CHI, nu2/CHI ) ;
+    return cv::Point2d( nu1/CHI, nu2/CHI ) ;
 }
 
 void PointMassLens::updateApparentAbs( ) {
